updating SQL and dem parser

+++ IxigoDemManager/src/main/java/com/ixigo/demmanager/models/database/Round_kill_eventsDao.java
@@ -1,5 +1,7 @@
 package com.ixigo.demmanager.models.database;
 
+import java.math.BigDecimal;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -61,11 +63,11 @@ public class Round_kill_eventsDao extends IxigoDao<Round_kill_eventsDto> {
 		this.dto.setIsfirstkill(isfirstkill);
 	}
 
-	public Long getEventtime() {
+	public BigDecimal getEventtime() {
 		return dto.getEventtime();
 	}
 
-	public void setEventtime(Long eventtime) {
+	public void setEventtime(BigDecimal eventtime) {
 		this.dto.setEventtime(eventtime);
 	}
