+++ Scripts/Docker/PostgreSql/initNewDemFilesDb.sql
@@ -77,7 +77,7 @@ CREATE TABLE PLAYER_ROUND_STATS (
 
 CREATE TABLE ROUND_EVENTS (
     eventType VARCHAR(255),
-    eventtime NUMERIC(6, 2), 
+    eventtime NUMERIC(10, 6), 
     steamID VARCHAR(255),
     round INTEGER,
     match_fileName VARCHAR(255),
@@ -85,7 +85,7 @@ CREATE TABLE ROUND_EVENTS (
 );
 
 CREATE TABLE ROUND_KILL_EVENTS (
-    eventtime INTEGER, 
+    eventtime NUMERIC(10, 6), 
     steamID VARCHAR(255),
     assister VARCHAR(255),
     flashAssister VARCHAR(255),
@@ -103,7 +103,7 @@ CREATE TABLE ROUND_KILL_EVENTS (
 
 CREATE TABLE ROUND_SHOT_EVENTS (
     eventType VARCHAR(255),
-    eventtime NUMERIC(6, 2), 
+    eventtime NUMERIC(10, 6), 
     steamID VARCHAR(255),
     round INTEGER,
     weapon VARCHAR(255),
@@ -112,7 +112,7 @@ CREATE TABLE ROUND_SHOT_EVENTS (
 );
 
 CREATE TABLE ROUND_HIT_EVENTS (
-    eventtime NUMERIC(6, 2), 
+    eventtime NUMERIC(10, 6), 
     steamID VARCHAR(255),
     round INTEGER,
     weapon VARCHAR(255),
